move space + project settings from route to button
