+++ apps/client/app/(navbar)/create/[id]/page.tsx
@@ -1,7 +1,89 @@
-import { fetchProject } from "../../../../src/server/helpers/fetchProject";
+import { Metadata } from "next";
+import Image from "next/image";
+import Link from "next/link";
+import { notFound } from "next/navigation";
 
-export default async function About({ params: { id } }: { params: { id: string } }) {
+import { fetchProject } from "../../../../src/server/helpers/fetchProject";
+import ButtonTabs, { TabContent } from "../../../../src/ui/ButtonTabs";
+import { isFromCDN } from "../../../../src/utils/isFromCDN";
+import About from "./About";
+import Settings from "./Settings";
+
+type Params = { id: string };
+
+export async function generateMetadata({ params: { id } }: { params: Params }): Promise<Metadata> {
   const project = await fetchProject(id);
 
-  return <div className="whitespace-pre-line">{project?.description}</div>;
+  if (!project) return {};
+
+  return {
+    title: project.name,
+    description: project.description,
+  };
+}
+
+interface Props {
+  params: Params;
+}
+
+export default async function Project({ params }: Props) {
+  const { id } = params;
+  const project = await fetchProject(id);
+
+  if (!project) notFound();
+
+  return (
+    <div className="flex justify-center">
+      <div className="max-w-content mx-4 space-y-8 py-8">
+        <div className="flex flex-col space-y-8 md:flex-row md:space-y-0 md:space-x-8">
+          <div className="aspect-card h-full w-full rounded-2xl bg-neutral-200">
+            <div className="relative h-full w-full object-cover">
+              {project.image &&
+                (isFromCDN(project.image) ? (
+                  <Image
+                    src={project.image}
+                    priority
+                    fill
+                    sizes="40vw"
+                    alt=""
+                    className="rounded-2xl object-cover"
+                  />
+                ) : (
+                  <img
+                    src={project.image}
+                    alt=""
+                    className="h-full w-full rounded-2xl object-cover"
+                    crossOrigin="anonymous"
+                  />
+                ))}
+            </div>
+          </div>
+
+          <div className="flex flex-col justify-between space-y-8 md:w-2/3">
+            <div className="space-y-4">
+              <div className="flex justify-center text-3xl font-black">{project?.name}</div>
+            </div>
+
+            <Link
+              href={`/editor/${id}`}
+              className="rounded-full bg-neutral-900 py-3 text-center text-lg font-bold text-white transition hover:scale-105 active:opacity-90"
+            >
+              Open Editor
+            </Link>
+          </div>
+        </div>
+
+        <ButtonTabs titles={["About", "Settings"]}>
+          <TabContent value="About">
+            {/* @ts-expect-error Server Component */}
+            <About params={params} />
+          </TabContent>
+          <TabContent value="Settings">
+            {/* @ts-expect-error Server Component */}
+            <Settings params={params} />
+          </TabContent>
+        </ButtonTabs>
+      </div>
+    </div>
+  );
 }
