+++ apps/client/app/(navbar)/space/[id]/Settings.tsx
@@ -1,8 +1,6 @@
-import { notFound } from "next/navigation";
-
-import { fetchSpace } from "../../../../../src/server/helpers/fetchSpace";
-import { getServerSession } from "../../../../../src/server/helpers/getServerSession";
-import RainbowkitWrapper from "../../../RainbowkitWrapper";
+import { fetchSpace } from "../../../../src/server/helpers/fetchSpace";
+import { getServerSession } from "../../../../src/server/helpers/getServerSession";
+import RainbowkitWrapper from "../../RainbowkitWrapper";
 import Delete from "./Delete";
 
 type Params = { id: string };
@@ -11,11 +9,11 @@ export default async function Settings({ params: { id } }: { params: Params }) {
   const spaceId = parseInt(id);
   const [session, space] = await Promise.all([getServerSession(), fetchSpace(spaceId)]);
 
-  if (!space || !session?.address) notFound();
+  if (!space || !session?.address) return null;
 
   const isOwner = session.address === space.owner;
 
-  if (!isOwner) notFound();
+  if (!isOwner) return null;
 
   return (
     <RainbowkitWrapper>
